// packages/schemas/src/signUpSchema.ts
import { z } from "zod";

export const signupSchema = z.object({
  name: z.string().min(1, "Name is required"),
  email: z.string().email("Invalid email address"),
  password: z.string().min(8, "Password must be at least 8 characters long"),
  number: z.string().max(10, "Invalid number").optional(),
});